+++ streaming_jupyter_integrations/magics.py
@@ -30,7 +30,7 @@ from pyflink.table import (EnvironmentSettings, ResultKind,
                            StreamTableEnvironment, Table, TableResult)
 
 from .cast_utils import cast_timestamp_ltz_to_string
-from .config_utils import load_config_file
+from .config_utils import load_config_file, read_flink_config_file
 from .deployment_bar import DeploymentBar
 from .display import pyflink_result_kind_to_string
 from .jar_handler import JarHandler
@@ -132,7 +132,7 @@ class Integrations(Magics):
         }
 
     def _flink_connect_local(self, port: int) -> None:
-        global_conf = self.__read_global_config()
+        global_conf = read_flink_config_file()
         conf = self.__create_configuration_from_dict(global_conf)
         conf.set_integer("rest.port", port)
         conf.set_integer("parallelism.default", 1)
@@ -143,11 +143,13 @@ class Integrations(Magics):
         )
 
     def _flink_connect_remote(self, hostname: str, port: int) -> None:
+        global_conf = read_flink_config_file()
+        conf = self.__create_configuration_from_dict(global_conf)
         gateway = get_gateway()
         empty_string_array = gateway.new_array(gateway.jvm.String, 0)
         self.s_env = StreamExecutionEnvironment(
             gateway.jvm.org.apache.flink.streaming.api.environment.StreamExecutionEnvironment.createRemoteEnvironment(
-                hostname, port, empty_string_array
+                hostname, port, conf._j_configuration, empty_string_array
             )
         )
 
@@ -636,19 +638,6 @@ class Integrations(Magics):
     @staticmethod
     def __retract_user_as_something_is_executing_in_background() -> None:
         print("Please wait for the previously submitted task to finish or cancel it.")
-
-    @staticmethod
-    def __read_global_config() -> Dict[str, Any]:
-        if "FLINK_HOME" in os.environ:
-            with open(os.path.join(os.environ["FLINK_HOME"], "conf", "flink-conf.yaml"), 'r') as stream:
-                try:
-                    parsed_yaml = yaml.safe_load(stream)
-                    return parsed_yaml
-                except yaml.YAMLError as exc:
-                    print(exc)
-        else:
-            print("FLINK_HOME environment variable is not set, reading flink-conf skipped")
-        return {}
 
     @staticmethod
     def __create_configuration_from_dict(new_values: Dict[str, Any]) -> Configuration: